Trying to deal with a macos-case-git issue by changing the formats directory to a different name and then back to a lower cased version.
Files checked in: DataRepo/Formats/DataRepo/FluxCircFormat.py DataRepo/Formats/DataRepo/PeakDataFormat.py DataRepo/Formats/DataRepo/PeakGroupsFormat.py DataRepo/Formats/DataRepo/SearchGroup.py DataRepo/Formats/DataRepo/__init__.py DataRepo/Formats/Format.py DataRepo/Formats/FormatGroup.py DataRepo/Formats/Query.py DataRepo/Formats/__init__.py DataRepo/forms.py DataRepo/templatetags/customtags.py DataRepo/tests/test_compositeviews.py DataRepo/tests/test_views.py DataRepo/views.py DataRepo/TMPFMT/DataRepo/FluxCircFormat.py DataRepo/TMPFMT/DataRepo/PeakDataFormat.py DataRepo/TMPFMT/DataRepo/PeakGroupsFormat.py DataRepo/TMPFMT/DataRepo/SearchGroup.py DataRepo/TMPFMT/DataRepo/__init__.py DataRepo/TMPFMT/Format.py DataRepo/TMPFMT/FormatGroup.py DataRepo/TMPFMT/Query.py DataRepo/TMPFMT/__init__.py
