+++ DataRepo/TMPFMT/DataRepo/PeakDataFormat.py
@@ -1,86 +1,181 @@
-from DataRepo.formats.Format import Format
-from DataRepo.models import Animal, PeakGroup, TracerLabeledClass
+from DataRepo.TMPFMT.Format import Format
+from DataRepo.TMPFMT.Query import (
+    appendFilterToGroup,
+    createFilterCondition,
+    createFilterGroup,
+)
+from DataRepo.models import Animal, PeakData, TracerLabeledClass
 
 
-class PeakGroupsFormat(Format):
+class PeakDataFormat(Format):
     """
     This class encapsulates all the metadata of a single search output format, which includes multiple tables/fields.
     """
 
-    id = "pgtemplate"
-    name = "PeakGroups"
-    rootmodel = PeakGroup
+    id = "pdtemplate"
+    name = "PeakData"
+    rootmodel = PeakData
     stats = [
         {
             "displayname": "Animals",
-            "distincts": ["msrun__sample__animal__name"],
+            "distincts": ["peak_group__msrun__sample__animal__name"],
             "filter": None,
         },
         {
-            "displayname": "Labeled Elements",
-            "distincts": ["peak_data__labeled_element"],
+            "displayname": "Labels",
+            "distincts": [
+                "labeled_element",
+                "labeled_count",
+            ],
+            "filter": None,
+            "delimiter": ":",
+        },
+        {
+            "displayname": "Feeding Statuses",
+            "distincts": ["peak_group__msrun__sample__animal__feeding_status"],
             "filter": None,
         },
         {
-            "displayname": "Measured Compounds",
-            "distincts": ["compounds__name"],
-            "filter": None,
+            "displayname": "Corrected Abundances",  # Append " > 0.1" based on filter
+            "distincts": ["corrected_abundance"],
+            "filter": appendFilterToGroup(
+                createFilterGroup(),
+                createFilterCondition("corrected_abundance", "gt", 0.1),
+            ),
         },
         {
             "displayname": "Samples",
-            "distincts": ["msrun__sample__name"],
+            "distincts": ["peak_group__msrun__sample__name"],
             "filter": None,
         },
         {
             "displayname": "Tissues",
-            "distincts": ["msrun__sample__tissue__name"],
+            "distincts": ["peak_group__msrun__sample__tissue__name"],
             "filter": None,
         },
         {
             "displayname": "Tracer Compounds",
-            "distincts": ["msrun__sample__animal__tracer_compound__name"],
+            "distincts": ["peak_group__msrun__sample__animal__tracer_compound__name"],
             "filter": None,
         },
         {
-            "displayname": "Studies",
-            "distincts": ["msrun__sample__animal__studies__name"],
-            "filter": None,
-        },
-        {
-            "displayname": "Feeding Statuses",
-            "distincts": ["msrun__sample__animal__feeding_status"],
-            "filter": None,
-        },
-        {
-            "displayname": "Infusion Rates",
-            "distincts": ["msrun__sample__animal__tracer_infusion_rate"],
-            "filter": None,
-        },
-        {
-            "displayname": "Infusion Concentrations",
-            "distincts": ["msrun__sample__animal__tracer_infusion_concentration"],
+            "displayname": "Measured Compounds",
+            "distincts": ["peak_group__compounds__name"],
             "filter": None,
         },
     ]
     model_instances = {
-        "PeakGroupSet": {
-            "model": "PeakGroupSet",
-            "path": "peak_group_set",
-            "reverse_path": "peak_groups",
+        "PeakData": {
+            "model": "PeakData",
+            "path": "",
+            "reverse_path": "",
             "manytomany": {
                 "is": False,
                 "split_rows": False,
             },
             "fields": {
                 "id": {
-                    "displayname": "(Internal) Peak Group Set Index",
+                    "displayname": "(Internal) Peak Data Index",
                     "searchable": True,
                     "displayed": False,  # Used in link
-                    "handoff": "filename",  # This is the field that will be loaded in the search form
+                    # "handoff": "",
+                    # Using in link will expose the internal index field in the search form because there's no
+                    # searchable unique field for handoff
                     "type": "number",
                 },
-                "filename": {
-                    "displayname": "Peak Group Set Filename",
+                "labeled_element": {
+                    "displayname": "Labeled Element",
+                    "searchable": True,
+                    "displayed": True,
+                    "type": "enumeration",
+                    "choices": TracerLabeledClass.TRACER_LABELED_ELEMENT_CHOICES,
+                },
+                "labeled_count": {
+                    "displayname": "Labeled Count",
+                    "searchable": True,
+                    "displayed": True,
+                    "type": "number",
+                },
+                "raw_abundance": {
+                    "displayname": "Raw Abundance",
+                    "searchable": True,
+                    "displayed": True,
+                    "type": "number",
+                },
+                "corrected_abundance": {
+                    "displayname": "Corrected Abundance",
+                    "searchable": True,
+                    "displayed": True,
+                    "type": "number",
+                },
+                "fraction": {
+                    "displayname": "Fraction",
+                    "searchable": False,  # Cannot search cached property
+                    "displayed": True,
+                    "type": "number",
+                },
+                "med_mz": {
+                    "displayname": "Median M/Z",
+                    "searchable": True,
+                    "displayed": True,
+                    "type": "number",
+                },
+                "med_rt": {
+                    "displayname": "Median RT",
+                    "searchable": True,
+                    "displayed": True,
+                    "type": "number",
+                },
+            },
+        },
+        "PeakGroup": {
+            "model": "PeakGroup",
+            "path": "peak_group",
+            "reverse_path": "peak_data",
+            "manytomany": {
+                "is": False,
+                "split_rows": False,
+            },
+            "fields": {
+                "id": {
+                    "displayname": "(Internal) Peak Group Index",  # Used in link
+                    "searchable": True,
+                    "displayed": False,
+                    "handoff": "name",  # This is the field that will be loaded in the search form
+                    "type": "number",
+                },
+                "name": {
+                    "displayname": "Peak Group",
+                    "searchable": True,
+                    "displayed": True,
+                    "type": "string",
+                },
+                "formula": {
+                    "displayname": "Formula",
+                    "searchable": True,
+                    "displayed": True,
+                    "type": "string",
+                },
+            },
+        },
+        "MeasuredCompound": {
+            "model": "Compound",
+            "path": "peak_group__compounds",
+            "reverse_path": "peak_groups__peak_data",
+            "manytomany": {
+                "is": False,
+                "split_rows": False,
+            },
+            "fields": {
+                "id": {
+                    "displayname": "(Internal) Measured Compound Index",
+                    "searchable": True,
+                    "displayed": False,  # Used in link
+                    "handoff": "name",  # This is the field that will be loaded in the search form
+                    "type": "number",
+                },
+                "name": {
+                    "displayname": "Measured Compound (Primary Synonym)",
                     "searchable": True,
                     "displayed": True,
                     "type": "string",
@@ -89,8 +184,8 @@ class PeakGroupsFormat(Format):
         },
         "CompoundSynonym": {
             "model": "CompoundSynonym",
-            "path": "compounds__synonyms",
-            "reverse_path": "compound__peak_groups",
+            "path": "peak_group__compounds__synonyms",
+            "reverse_path": "compound__peak_groups__peak_data",
             "manytomany": {
                 "is": True,
                 "split_rows": False,
@@ -111,55 +206,146 @@ class PeakGroupsFormat(Format):
                 },
             },
         },
-        "PeakGroup": {
-            "model": "PeakGroup",
-            "path": "",
-            "reverse_path": "",
+        "PeakGroupSet": {
+            "model": "PeakGroupSet",
+            "path": "peak_group__peak_group_set",
+            "reverse_path": "peak_groups__peak_data",
             "manytomany": {
                 "is": False,
                 "split_rows": False,
             },
             "fields": {
                 "id": {
-                    "displayname": "(Internal) Peak Group Index",
+                    "displayname": "(Internal) Peak Group Set Index",  # Used in link
+                    "searchable": True,
+                    "displayed": False,
+                    "handoff": "filename",  # This is the field that will be loaded in the search form
+                    "type": "number",
+                },
+                "filename": {
+                    "displayname": "Peak Group Set Filename",
+                    "searchable": True,
+                    "displayed": True,
+                    "type": "string",
+                },
+            },
+        },
+        "Sample": {
+            "model": "Sample",
+            "path": "peak_group__msrun__sample",
+            "reverse_path": "msruns__peak_groups__peak_data",
+            "manytomany": {
+                "is": False,
+                "split_rows": False,
+            },
+            "fields": {
+                "id": {
+                    "displayname": "(Internal) Sample Index",  # Used in link
+                    "searchable": True,
+                    "displayed": False,
+                    "type": "number",
+                    "handoff": "name",
+                },
+                "name": {
+                    "displayname": "Sample",
+                    "searchable": True,
+                    "displayed": True,
+                    "type": "string",
+                },
+            },
+        },
+        "Tissue": {
+            "model": "Tissue",
+            "path": "peak_group__msrun__sample__tissue",
+            "reverse_path": "samples__msruns__peak_groups__peak_data",
+            "manytomany": {
+                "is": False,
+                "split_rows": False,
+            },
+            "fields": {
+                "id": {
+                    "displayname": "Tissue",
+                    "searchable": True,
+                    "displayed": False,
+                    "type": "number",
+                    "handoff": "name",
+                },
+                "name": {
+                    "displayname": "Tissue",
+                    "searchable": True,
+                    "displayed": True,
+                    "type": "string",
+                },
+            },
+        },
+        "Animal": {
+            "model": "Animal",
+            "path": "peak_group__msrun__sample__animal",
+            "reverse_path": "samples__msruns__peak_groups__peak_data",
+            "manytomany": {
+                "is": False,
+                "split_rows": False,
+            },
+            "fields": {
+                "id": {
+                    "displayname": "(Internal) Animal Index",
                     "searchable": True,
                     "displayed": False,  # Used in link
                     "handoff": "name",  # This is the field that will be loaded in the search form
                     "type": "number",
                 },
                 "name": {
-                    "displayname": "Peak Group",
+                    "displayname": "Animal",
                     "searchable": True,
                     "displayed": True,
                     "type": "string",
                 },
-                "formula": {
-                    "displayname": "Formula",
+                "body_weight": {
+                    "displayname": "Body Weight (g)",
+                    "searchable": True,
+                    "displayed": True,
+                    "type": "number",
+                },
+                "genotype": {
+                    "displayname": "Genotype",
                     "searchable": True,
                     "displayed": True,
                     "type": "string",
                 },
-                "enrichment_fraction": {
-                    "displayname": "Enrichment Fraction",
-                    "searchable": False,  # Cannot search cached property
+                "sex": {
+                    "displayname": "Sex",
+                    "searchable": True,
+                    "displayed": True,
+                    "type": "enumeration",
+                    "choices": Animal.SEX_CHOICES,
+                },
+                "age": {
+                    "displayname": "Age (d-hh:mm:ss)",
+                    "searchable": False,
                     "displayed": True,
                     "type": "number",
                 },
-                "enrichment_abundance": {
-                    "displayname": "Enrichment Abundance",
-                    "searchable": False,  # Cannot search cached property
+                "feeding_status": {
+                    "displayname": "Feeding Status",
+                    "searchable": True,
+                    "displayed": True,
+                    "type": "string",
+                },
+                "diet": {
+                    "displayname": "Diet",
+                    "searchable": True,
+                    "displayed": True,
+                    "type": "string",
+                },
+                "tracer_infusion_concentration": {
+                    "displayname": "Tracer Infusion Concentration (mM)",
+                    "searchable": True,
                     "displayed": True,
                     "type": "number",
                 },
-                "total_abundance": {
-                    "displayname": "Total Abundance",
-                    "searchable": False,  # Cannot search cached property
-                    "displayed": True,
-                    "type": "number",
-                },
-                "normalized_labeling": {
-                    "displayname": "Normalized Labeling",
-                    "searchable": False,  # Cannot search cached property
+                "tracer_infusion_rate": {
+                    "displayname": "Tracer Infusion Rate (ul/min/g)",
+                    "searchable": True,
                     "displayed": True,
                     "type": "number",
                 },
@@ -167,8 +353,8 @@ class PeakGroupsFormat(Format):
         },
         "Protocol": {
             "model": "Protocol",
-            "path": "msrun__sample__animal__treatment",
-            "reverse_path": "animals__samples__msruns__peak_groups",
+            "path": "peak_group__msrun__sample__animal__treatment",
+            "reverse_path": "animals__samples__msruns__peak_groups__peak_data",
             "manytomany": {
                 "is": False,
                 "split_rows": False,
@@ -189,138 +375,10 @@ class PeakGroupsFormat(Format):
                 },
             },
         },
-        "Sample": {
-            "model": "Sample",
-            "path": "msrun__sample",
-            "reverse_path": "msruns__peak_groups",
-            "manytomany": {
-                "is": False,
-                "split_rows": False,
-            },
-            "fields": {
-                "id": {
-                    "displayname": "(Internal) Sample Index",
-                    "searchable": True,
-                    "displayed": False,  # Used in link
-                    "handoff": "name",  # This is the field that will be loaded in the search form
-                    "type": "number",
-                },
-                "name": {
-                    "displayname": "Sample",
-                    "searchable": True,
-                    "displayed": True,
-                    "type": "string",
-                },
-            },
-        },
-        "Tissue": {
-            "model": "Tissue",
-            "path": "msrun__sample__tissue",
-            "reverse_path": "samples__msruns__peak_groups",
-            "manytomany": {
-                "is": False,
-                "split_rows": False,
-            },
-            "fields": {
-                "id": {
-                    "displayname": "(Internal) Tissue Index",
-                    "searchable": True,
-                    "displayed": False,  # Used in link
-                    "handoff": "name",  # This is the field that will be loaded in the search form
-                    "type": "number",
-                },
-                "name": {
-                    "displayname": "Tissue",
-                    "searchable": True,
-                    "displayed": True,
-                    "type": "string",
-                },
-            },
-        },
-        "Animal": {
-            "model": "Animal",
-            "path": "msrun__sample__animal",
-            "reverse_path": "samples__msruns__peak_groups",
-            "manytomany": {
-                "is": False,
-                "split_rows": False,
-            },
-            "fields": {
-                "id": {
-                    "displayname": "(Internal) Animal Index",
-                    "searchable": True,
-                    "displayed": False,  # Used in link
-                    "handoff": "name",  # This is the field that will be loaded in the search form
-                    "type": "number",
-                },
-                "name": {
-                    "displayname": "Animal",
-                    "searchable": True,
-                    "displayed": True,
-                    "type": "string",
-                },
-                "genotype": {
-                    "displayname": "Genotype",
-                    "searchable": True,
-                    "displayed": True,
-                    "type": "string",
-                },
-                "body_weight": {
-                    "displayname": "Body Weight (g)",
-                    "searchable": True,
-                    "displayed": True,
-                    "type": "number",
-                },
-                "age": {
-                    "displayname": "Age (d-hh:mm:ss)",
-                    "searchable": False,  # currently no data available for testing and a bug: issue #180
-                    "displayed": True,
-                    "type": "number",
-                },
-                "sex": {
-                    "displayname": "Sex",
-                    "searchable": True,
-                    "displayed": True,
-                    "type": "enumeration",
-                    "choices": Animal.SEX_CHOICES,
-                },
-                "diet": {
-                    "displayname": "Diet",
-                    "searchable": True,
-                    "displayed": True,
-                    "type": "string",
-                },
-                "feeding_status": {
-                    "displayname": "Feeding Status",
-                    "searchable": True,
-                    "displayed": True,
-                    "type": "string",
-                },
-                "tracer_labeled_atom": {
-                    "displayname": "Tracer Labeled Element",
-                    "searchable": True,
-                    "displayed": True,
-                    "type": "enumeration",
-                    "choices": TracerLabeledClass.TRACER_LABELED_ELEMENT_CHOICES,
-                },
-                "tracer_infusion_rate": {
-                    "displayname": "Tracer Infusion Rate (ul/min/g)",
-                    "searchable": True,
-                    "displayed": True,
-                    "type": "number",
-                },
-                "tracer_infusion_concentration": {
-                    "displayname": "Tracer Infusion Concentration (mM)",
-                    "searchable": True,
-                    "displayed": True,
-                    "type": "number",
-                },
-            },
-        },
         "TracerCompound": {
             "model": "Compound",
-            "path": "msrun__sample__animal__tracer_compound",
-            "reverse_path": "animals__samples__msruns__peak_groups",
+            "path": "peak_group__msrun__sample__animal__tracer_compound",
+            "reverse_path": "animals__samples__msruns__peak_groups__peak_data",
             "manytomany": {
                 "is": False,
                 "split_rows": False,
@@ -341,35 +399,10 @@ class PeakGroupsFormat(Format):
                 },
             },
         },
-        "MeasuredCompound": {
-            "model": "Compound",
-            "path": "compounds",
-            "reverse_path": "peak_groups",
-            "manytomany": {
-                "is": True,
-                "split_rows": False,
-                "root_annot_fld": "compound",
-            },
-            "fields": {
-                "id": {
-                    "displayname": "(Internal) Measured Compound Index",
-                    "searchable": True,
-                    "displayed": False,  # Used in link
-                    "handoff": "name",  # This is the field that will be loaded in the search form
-                    "type": "number",
-                },
-                "name": {
-                    "displayname": "Measured Compound (Primary Synonym)",
-                    "searchable": True,
-                    "displayed": True,  # Will display due to the handoff
-                    "type": "string",
-                },
-            },
-        },
         "Study": {
             "model": "Study",
-            "path": "msrun__sample__animal__studies",
-            "reverse_path": "animals__samples__msruns__peak_groups",
+            "path": "peak_group__msrun__sample__animal__studies",
+            "reverse_path": "animals__samples__msruns__peak_groups__peak_data",
             "manytomany": {
                 "is": True,
                 "split_rows": False,
